backend/requirements.txt
Django==4.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
gunicorn==21.2.0
mysqlclient==2.1.1
pyjwt==2.8.0
python-dotenv==0.21.0
